+++ feature/search/src/main/java/com/wepli/search/detail/viewmodel/SearchDetailViewModel.kt
@@ -2,9 +2,9 @@ package com.wepli.search.detail.viewmodel
 
 import base.BaseMviViewModel
 import com.wepli.core.kotlin.suspendCollectResult
-import com.wepli.search.detail.state.SearchDetailEffect
-import com.wepli.search.detail.state.SearchDetailIntent
-import com.wepli.search.detail.state.SearchDetailUiState
+import com.wepli.search.detail.mvi.SearchDetailEffect
+import com.wepli.search.detail.mvi.SearchDetailIntent
+import com.wepli.search.detail.mvi.SearchDetailUiState
 import com.wepli.uimodel.music.SongUiData
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
